requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.1
constantly==23.10.4
contourpy==1.1.1
cryptography==41.0.5
cycler==0.12.1
daphne==4.0.0
Django==4.2.6
django-bulk-update-or-create==0.3.0
django-cors-headers==4.3.0
django-cron==0.6.0
django-debug-toolbar==4.2.0
django-filter==23.3
django-pandas==0.6.6
django-polymorphic==3.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
fcm-django==2.0.0
firebase-admin==6.2.0
fonttools==4.43.1
google-api-core==2.12.0
google-api-python-client==2.105.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.0
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.0
grpcio-status==1.59.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kiwisolver==1.4.5
matplotlib==3.8.0
msgpack==1.0.7
namegenerator==1.0.6
numpy==1.26.1
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
proto-plus==1.22.3
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
zope.interface==6.1
